+++ src/routes/userRouter.js
@@ -1,9 +1,9 @@
 import express from 'express';
-import {asyncWrapper, authenticate} from '../../utils/middleware.js';
+import { asyncWrapper, authenticate } from '../../utils/middleware.js';
 import { getUserProfile } from '../controllers/user/getUserProfile.js';
 import { kakaoLogin, kakaoCallback } from '../config/kakaoStrategy.js';
 import { alreadyExists } from '../modules/user.module.js';
-import jwtService from '../modules/jwt.module.js';
+// import jwtService from '../modules/jwt.module.js';
 // import { googleLoginCheck } from '../passport/googleStrategy.js';
 // import { googleCallback } from '../passport/googleStrategy.js';
 import {
@@ -11,7 +11,9 @@ import {
     signupForCompanyController,
     signinForCompanyController,
     validateEmailController,
-} from '../controllers/user/companySignup.js';
+    verifyMissingPasswordToEmailController,
+    setNewPasswordController,
+} from '../controllers/user/uwreckcarAccount.js';
 import Slack from '../../config/slackbot.config.js';
 const router = express.Router();
 
@@ -31,7 +33,7 @@ router.get('/api/auth/kakao/callback', kakaoCallback, async (req, res) => {
         res.status(200).send({ access_token: access_token, refresh_token: refresh_token });
     } catch (err) {
         console.error(err);
-        await Slack('KakaoLogin', err)
+        await Slack('KakaoLogin', err);
         res.status(500).send({ errorMessage: err.message, stack: err.stack });
     }
 });
@@ -42,6 +44,8 @@ router.post('/api/users/signup', asyncWrapper(signupForCompanyController));
 router.post('/api/users/login', asyncWrapper(signinForCompanyController));
 router.post('/api/users/email', asyncWrapper(sendEmailController));
 router.post('/api/users/emailverify', asyncWrapper(validateEmailController));
+router.post('/api/users/passwordverify', asyncWrapper(verifyMissingPasswordToEmailController));
+router.post('/api/users/setnewpassword', asyncWrapper(setNewPasswordController));
 
 // router.get('/api/auth/google', googleLoginCheck); // 프로파일과 이메일 정보를 받는다.
 // // 위에서 구글 서버 로그인이 되면, 구글 redirect url 설정에 따라 이쪽 라우터로 오게 된다. 인증 코드를 박게됨
